refactor: 위젯, 파일명 변경
