+++ lib/core/get_responsive_ratio_by_width.dart
